backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
aiofiles==24.1.0
ultralytics==8.3.56
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
pydantic==2.10.4
